compiled version

+++ Rooms.h
@@ -42,7 +42,7 @@ int getRoomDifficulty(Room room);
 roomResult checkRoomParameters(char *email, int id, int price, int num_ppl,
                                int time_start, int time_finish, int difficulty);
 
-double getRoomRecommendedCalculation(Room room, int P_e, int skill_level);
+int getRoomRecommendedCalculation(Room room, int P_e, int skill_level);
 
 roomResult roomCheckIfOpen(Room room, int req_hour);
